+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import GlobalStyle from './GlobalStyle';
-import HomePage from './pages/HomePage';
+import MainPage from './pages/MainPage';
 import NotFoundPage from './pages/NotFoundPage';
 import List from './pages/List';
 
@@ -10,10 +10,10 @@ function App() {
       <GlobalStyle />
       <Routes>
         <Route path="/">
-          <Route index element={<HomePage />} />
+          <Route index element={<MainPage />} />
           <Route path="/list" element={<List />} />
-          <Route path="/post/$id" element={<HomePage />} />
-          <Route path="/post/$id/answer" element={<HomePage />} />
+          <Route path="/post/$id" element={<MainPage />} />
+          <Route path="/post/$id/answer" element={<MainPage />} />
         </Route>
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
