+++ src/pages/List/index.js
@@ -31,9 +31,8 @@ const List = () => {
     fetchData();
   }, []);
 
-  function handleSorting(e) {
+  const handleSorting = e => {
     // 정렬기능
-    setSortBy(e.target.innerHTML);
     const selectedSort = e.target.innerHTML;
     if (selectedSort === '이름순') {
       setCardInfo(
@@ -48,29 +47,29 @@ const List = () => {
     }
 
     setSortBy(selectedSort);
-  }
+  };
 
-  function handlePageChange(e) {
+  const handlePageChange = e => {
     // 페이지이동
     setCurrentPage(Number(e.target.innerHTML));
     setTranslateY((Number(e.target.innerHTML) - 1) * -414);
-  }
+  };
 
-  function goToPrev() {
+  const goToPrev = () => {
     // 이전 페이지리스트
     if (currentScroll > 1) {
       setTranslateX(translateX + 200);
       setCurrentScroll(currentScroll - 1);
     }
-  }
+  };
 
-  function goToNext() {
+  const goToNext = () => {
     // 다음 페이지리스트
     if (currentScroll < Math.ceil(totalPages / 5)) {
       setTranslateX(translateX - 200);
       setCurrentScroll(currentScroll + 1);
     }
-  }
+  };
 
   return (
     <Container x={translateX} y={translateY}>
@@ -96,7 +95,7 @@ const List = () => {
                 ? `/assets/images/arrowDown.svg`
                 : `/assets/images/arrowUp.svg`
             }
-            alt="arrow"
+            alt="정렬버튼 화살표"
           />
           {isDropDown && (
             <div className="list-main-ul-dropdown">
@@ -129,7 +128,7 @@ const List = () => {
                 <div className="card-profile">
                   <img
                     src={item.imageSource}
-                    alt="profile"
+                    alt="프로필사진"
                     className="card-profile-img"
                   />
                   <h2 className="card-profile-h2">{item.name}</h2>
@@ -138,7 +137,7 @@ const List = () => {
                   <div className="card-questions-div">
                     <img
                       src="/assets/images/messages.svg"
-                      alt="message"
+                      alt="메세지아이콘"
                       className="card-questions-img"
                     />
                     <h3 className="card-questions-h3">받은질문</h3>
